members move into bizLicAppManager
